+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/sync/SupporterCancelRequest.kt
@@ -1,9 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.servers.sync
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class SupporterCancelRequest(
-    @field:Json(name = "bundleUuid") val bundleUuid: String
-)